Tweak templates in services to render the right thing.
This is annoyingly complicated because `services` as a directory is
a Hugo 'section', so the index page expects to be a list. To override
the default behaviour and render the data template, we need to rename
the files to _index.{en,cs}.md and override the default _list layout
by setting 'layout: single' in the front matter. After that the memogis
page is rendered just fine.